+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Added
 - setting for setup unknown author display name
 - setting for sending notifications by email
-- link in IMPORTRANGE support
+- support IMPORTRANGE
 
 ## Changed
 - fix address when opening editor