+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix address when opening editor
 - URL for requests to Conversion API
 - reading JSON instead of XML from Conversion API
+- Nextcloud v28 is no longer supported
 
 ## 9.5.0
 ## Added